requirements.txt
Django==3.1
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cron-descriptor==1.2.24
croniter==0.3.31
django-appconf==1.0.4
django-compressor==2.4
django-environ==0.4.5
django-redis==4.12.1
idna==2.10
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2020.1
rcssmin==1.0.6
redis==3.5.3
requests==2.23.0
rjsmin==1.1.0
sentry-sdk==0.17.0
six==1.15.0
sqlparse==0.3.1
statsd==3.3.0
urllib3==1.25.10
whitenoise==5.2.0
